Add IntervalMixin, use in {elastic,rabbitmq}-stats.py
* rabbitmq-stats.py: eliminiate startup loop, catch all loop errors
* elastic-stats.py: add BYTES for clarity, commenting
